nvidia-detect: pass __u16 argument to function nv_lookup_device_id()
Function nv_lookup_device_id() expects an int as it's argument but
receives dev->device_id which is __u16. Making the pci_ids[] arrays
unsigned short and passing __u16 to nv_lookup_device_id() saves 1.8K on
the compiled program size (64-bit).

+++ nvidia-detect/nvidia-detect.c
@@ -133,7 +133,7 @@ static void list_all_nvidia_devices(void)
 	}
 }
 
-static int nv_lookup_device_id(int device_id)
+static int nv_lookup_device_id(u_int16_t device_id)
 {
 	size_t i;
